requirements.txt
streamlit==1.32.0
python-dotenv==1.0.0
moviepy==1.0.3
openai==1.12.0
requests==2.31.0
firebase-admin==6.2.0
imageio==2.34.0
imageio-ffmpeg==0.4.9
decorator==4.4.2  # Changed to be compatible with moviepy
tqdm==4.66.2
numpy==1.26.4